Add files via upload
This script is using machine learning to classify tweets as either positive or negative based on the text of the tweet. The script starts by importing several libraries that will be used later in the script, including pandas, numpy, and nltk. It then downloads a set of "stopwords" (common words that don't provide much information) from nltk. The script reads in two data files, "NLP_train.csv" and "NLP_test.csv", and combines them into one data frame.

The script then performs some data cleaning by dropping unnecessary columns and resetting the index. It also checks for missing values in the 'OriginalTweet' column.

The script then starts the preprocessing of the tweets by creating an empty list called 'corpus'. It then loops through all the tweets, removing any non-alphabetic characters, converting the tweets to lowercase, tokenizing the tweets, removing stopwords, and stemming the words. It then rejoins the processed words and appends the tweet to the 'corpus' list.

The script then initializes the CountVectorizer() class and uses it to generate a feature matrix and target variable. Then the script starts the classification process by training different classifiers, including Decision Tree, KNN, Random Forest, Logistic Regression and Naive Bayes on the feature matrix. It then tests these classifiers on the test dataset and prints the confusion matrix and accuracy score for each classifier.